+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## unreleased
+## 4.9.0
 * Add `paypal_messages` to `Dispute`
 * Add `tax_identifiers` parameter to `Customer.create` and `Customer.update`